people module: Move refactored pm_perma_link to interal_url.js.
In this commit, refactored pm_perma_link was moved from people.js
to internal_url.js to be shared with the mobile app in the future.
In people.js was created a wrapper function for pm_perma_link.
